Move commit components to component directory
