+++ app/components/commitfunds/Aftermath.scss
@@ -1,4 +1,4 @@
-@import '../../common/sass/variables';
+@import '../../../common/sass/variables';
 
 .aftermath {
   .info-box {
