+++ app/components/commitfunds/CommitFundsDescription.scss
@@ -1,4 +1,4 @@
-@import '../../common/sass/variables';
+@import '../../../common/sass/variables';
 
 .container {
   font-size: $font-size-smallest;
