+++ app/components/commitfunds/CommitFundsForm.scss
@@ -1,4 +1,4 @@
-@import '../../common/sass/variables';
+@import '../../../common/sass/variables';
 
 .form-container {
   position: relative;
