+++ app/components/commitfunds/CommitFundsStatic.scss
@@ -1,4 +1,4 @@
-@import '../../common/sass/variables';
+@import '../../../common/sass/variables';
 
 $spacing-from-container: 22px;
