+++ app/components/commitfunds/CommitHeaderComponent.scss
@@ -1,4 +1,4 @@
-@import '../../common/sass/variables';
+@import '../../../common/sass/variables';
 
 .header {
   position: relative;
