+++ app/components/commitfunds/CommitNavbar.scss
@@ -1,9 +1,9 @@
-@import '../../common/sass/variables';
+@import '../../../common/sass/variables';
 
 .brand {
   margin-top: 10px;
   border-right: 1px solid $white;
-  background: url('../../page/assets/NeuFund_icon_white.png');
+  background: url('../../../page/assets/NeuFund_icon_white.png');
   background-repeat: no-repeat;
   background-size: 30px;
   padding: 0;
